dataflow/requirements.txt
apache-beam[gcp]==2.17.0
cython
google-cloud-language
google-api-core